Rename namespaces and directories for integration tests
